udpated fit_sky_residuals
